+++ h2gis-functions/src/main/java/org/h2gis/functions/io/fgb/FGBRead.java
@@ -22,26 +22,27 @@ package org.h2gis.functions.io.fgb;
 import org.h2gis.api.AbstractFunction;
 import org.h2gis.api.EmptyProgressVisitor;
 import org.h2gis.api.ScalarFunction;
+import org.h2gis.functions.io.fgb.fileTable.FGBDriver;
 import org.h2gis.utilities.URIUtilities;
 
+import java.io.File;
 import java.io.IOException;
 import java.sql.Connection;
 import java.sql.SQLException;
 
 /**
- * Expose FlatGeobuffer writer function
+ * Expose FlatGeobuffer reader function
  * @author Erwan Bocher
  * @author Nicolas Fortin
  */
-public class FGBWrite extends AbstractFunction implements ScalarFunction {
+public class FGBRead extends AbstractFunction implements ScalarFunction {
 
-    public FGBWrite() {
-        addProperty(PROP_REMARKS, "Export a spatial table to a FlatGeobuf file.\n "
-                + "\nFGBWrite(..."
+    public FGBRead() {
+        addProperty(PROP_REMARKS, "Import a a FlatGeobuf file into a spatial table.\n "
+                + "\nFGBRead(..."
                 + "\n Supported arguments :"
                 + "\n path of the file, table name"
-                + "\n path of the file, table name, true to delete the file if exists"
-                + "\n path of the file, table name, true to delete the file if exists,options default 'createIndex=true nodeSize=16'");
+                + "\n path of the file, table name, true to delete the table if exists");
     }
 
     @Override
@@ -49,10 +50,6 @@ public class FGBWrite extends AbstractFunction implements ScalarFunction {
         return "execute";
     }
 
-    public static void execute(Connection connection, String fileName, String tableReference, boolean deleteFile, String option) throws SQLException, IOException {
-        FGBDriverFunction geobufDriverFunction = new FGBDriverFunction();
-        geobufDriverFunction.exportTable(connection, tableReference, URIUtilities.fileFromString(fileName), option ,deleteFile, new EmptyProgressVisitor());
-    }
     /**
      * Read a table and write it into a FlatGeobuf file.
      *
@@ -60,12 +57,13 @@ public class FGBWrite extends AbstractFunction implements ScalarFunction {
      * @param fileName       FlatGeobuf file name or URI
      * @param tableReference Table name or select query Note : The select query
      *                       must be enclosed in parenthesis
-     * @param deleteFile     true to delete output file
      * @throws IOException
      * @throws SQLException
      */
-    public static void execute(Connection connection, String fileName, String tableReference, boolean deleteFile) throws SQLException, IOException {
-        execute(connection, fileName, tableReference, deleteFile, "");
+    public static void execute(Connection connection, String fileName, String tableReference, boolean deleteTable) throws SQLException, IOException {
+        File file = URIUtilities.fileFromString(fileName);
+        FGBDriverFunction driver = new FGBDriverFunction();
+        driver.importFile(connection, tableReference, file, deleteTable ,new EmptyProgressVisitor());
     }
 
     /**